requirements.txt
Flask
flask-cors
python-docx
streamlit
PyPDF2
requests
python-dotenv
